refactor some dmctl package
