+++ dm/ctl/master/sql_replace.go
@@ -44,7 +44,7 @@ func sqlReplaceFunc(cmd *cobra.Command, _ []string) {
 		return
 	}
 
-	binlogPos, sqlPattern, sharding, err := extractBinlogPosSQLPattern(cmd)
+	binlogPos, sqlPattern, sharding, err := common.ExtractBinlogPosSQLPattern(cmd)
 	if err != nil {
 		common.PrintLines("%s", err.Error())
 		return
